# output/vietcms/architecture/openapi-v1.yaml
openapi: "3.1.0"
info:
  title: VietCMS API
  version: "1.0.0"
  description: >
    REST API cho VietCMS — no-code Marketing CMS SaaS multi-tenant.
    Tenant suy ra từ JWT claim `tenant_id`. Lỗi theo RFC 7807 (application/problem+json).
    Xem 04-api-design-v1.md để biết conventions & error catalog.
servers:
  - url: https://api.vietcms.vn/v1
    description: Production
  - url: https://staging-api.vietcms.vn/v1
    description: Staging
security:
  - BearerAuth: []
...
paths:
...
  /sites/{site_id}/contents:
    parameters:
      - $ref: "#/components/parameters/SiteId"
    get:
      summary: Danh sách trang/bài (US-01)
      operationId: listContents
      tags: [Content]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - { name: kind, in: query, schema: { $ref: "#/components/schemas/ContentKind" } }
        - { name: status, in: query, schema: { $ref: "#/components/schemas/ContentStatus" } }
        - { name: q, in: query, schema: { type: string }, description: "Tìm theo tiêu đề" }
        - { name: sort, in: query, schema: { type: string, default: "-updated_at" } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Content" } }
                  pagination: { $ref: "#/components/schemas/CursorPage" }
    post:
      summary: Tạo nội dung
      operationId: createContent
      tags: [Content]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, title]
              properties:
                kind: { $ref: "#/components/schemas/ContentKind" }
                title: { type: string, maxLength: 200 }
                slug: { type: string }
                template_id: { type: [string, "null"], format: uuid }
      responses:
        "201":
          description: Đã tạo (draft)
          content: { application/json: { schema: { $ref: "#/components/schemas/Content" } } }
        "409": { $ref: "#/components/responses/SlugConflict" }
...
components:
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
      description: Con trỏ phân trang (opaque)
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 20, maximum: 100, minimum: 1 }
...
    SiteId:
      name: site_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    ContentKind: { type: string, enum: [page, post] }
    ContentStatus: { type: string, enum: [draft, review, approved, scheduled, published, archived] }
...
    Content:
      type: object
      required: [id, site_id, kind, title, slug, status, updated_at]
      properties:
        id: { type: string, format: uuid, readOnly: true }
        site_id: { type: string, format: uuid }
        kind: { $ref: "#/components/schemas/ContentKind" }
        title: { type: string, maxLength: 200 }
        slug: { type: string }
        status: { $ref: "#/components/schemas/ContentStatus" }
        author: { $ref: "#/components/schemas/UserRef" }
        seo_score: { type: [integer, "null"] }
        scheduled_at: { type: [string, "null"], format: date-time }
        updated_at: { type: string, format: date-time, readOnly: true }
...
    UserRef:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
...
    CursorPage:
      type: object
      required: [next_cursor, has_more]
      properties:
        next_cursor: { type: [string, "null"] }
        has_more: { type: boolean }
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
  responses:
...
    SlugConflict:
      description: Slug trùng trong site (BR-02)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
          example:
            type: https://api.vietcms.vn/errors/slug-conflict
            title: Slug conflict
            status: 409
            detail: Đường dẫn đã tồn tại trong site.
            errors: [{ field: slug, message: "Chọn slug khác." }]
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT